backend/pyproject.toml
[tool.poetry]
name = "discoveredcheck"
version = "0.1.0"
description = "Chess analytics platform focused on human-centered analysis"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
django-cors-headers = "^4.4"
django-environ = "^0.11"
psycopg2-binary = "^2.9"
python-chess = "^1.10"
requests = "^2.32"
pillow = "^11.0"
whitenoise = "^6.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-django = "^4.9"
black = "^24.10"
ruff = "^0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
